InteractionEngineCpp to InteractionEngine
