+++ simfection/interaction_engine.py
@@ -1,9 +1,11 @@
 import pandas as pd
 import numpy as np
+from invoke import run
 # User defined imports
-from .connection_engine import ConnectionEngine
+import network
 from .settings import SimfectionSettings
 from .logger import SimfectionLogger
+from network import PyInteractions
 
 simfection_logger = SimfectionLogger(name=__name__)
 logger = simfection_logger.get_logger()
@@ -26,7 +28,7 @@ class InteractionEngine():
     def _init_interactions(self):
         # Initialize the network.PyInteractions object
         # For a look at the API, checkout the network.pyx file in /cpp_src/cython
-        self.interactions = network.PyInteractions()
+        interactions = network.PyInteractions()
 
         # Get all the population information required for running PyInteractions
         pop_agent = self.population['agent'].tolist()
@@ -40,7 +42,7 @@ class InteractionEngine():
         pop_immunity = self.population['immunity'].tolist()
 
         # Set the Population in the PyInteractions object
-        self.interactions.set_population(pop_agent,
+        interactions.set_population(pop_agent,
                                     pop_state,
                                     pop_infected_by,
                                     pop_days_infected,
@@ -53,7 +55,7 @@ class InteractionEngine():
         conn_max_connections = self.connections['max_connections'].tolist()
 
         # Set the Connections in the PyInteractions object
-        self.interactions.set_connections(conn_agent,
+        interactions.set_connections(conn_agent,
                                      conn_connections,
                                      conn_num_connections,
                                      conn_max_connections)
@@ -61,147 +63,111 @@ class InteractionEngine():
         # Set the Pathogen settings, first convert each value in pathogen
         # to unicode bytes for ingestion to interactions
         pathogen = {key.encode(): value for key, value in self.pathogen.items()}
-        self.interactions.set_pathogen_settings(pathogen)
+        interactions.set_pathogen_settings(pathogen)
 
-    def _make_interaction_pair(self, row):
-        agent_a = row['agent_a']
-        agent_b = row['agent_b']
-
-        if agent_a < agent_b:
-            return (row['agent_a'], row['agent_b'])
-        else:
-            return (row['agent_b'], row['agent_a'])
-
-    def _get_unique_connections(self):
-        connections = self.connections
-
-        # Explode on connections
-        # Drop num_connections column
-        # Rename columns
-        connections = (connections.query('num_connections > 0')
-                       .explode(column='connections')
-                       .drop(columns=['num_connections'])
-                       .rename(columns={
-                           'agent': 'agent_a',
-                           'connections': 'agent_b'
-                       })
-                       )
-
-        # Combine into interaction pair
-        connections['interaction'] = (connections
-                                      .apply(self._make_interaction_pair, axis=1)
-                                      )
-
-        # Drop unneeded columns
-        # Drop duplicates
-        connections = (connections
-                       .filter(['interaction'])
-                       .drop_duplicates()
-                       )
-
-        self.connections = connections
-
-    def _get_agents(self, a, b):
-        population = self.population
-
-        records = population.query('agent in {}'.format([a, b])).to_dict()
-        agent_a = {
-            'agent': records['agent'][a],
-            'state': records['state'][a]
-        }
-        agent_b = {
-            'agent': records['agent'][b],
-            'state': records['state'][b]
-        }
-
-        return agent_a, agent_b
-
-    def _qualify_interaction(self, agent_a, agent_b):
-
-        # Determine if agents have different states
-        different_states = agent_a['state'] != agent_b['state']
-
-        if different_states:
-
-            # Determine if someone is infected
-            someone_infected = agent_a['state'] == 'inf' or agent_b['state'] == 'inf'
-
-            # Determine if someone is susceptible
-            someone_susceptible = agent_a['state'] == 'sus' or agent_b['state'] == 'sus'
-
-            if someone_infected and someone_susceptible:
-
-                # Determine who is infected and who is susceptible
-                if agent_a['state'] == 'inf':
-                    infected = agent_a
-                    susceptible = agent_b
-                else:
-                    infected = agent_b
-                    susceptible = agent_a
-
-                # Determine if infected is contagious
-                days_infected = (
-                    self.population
-                    .query('agent == {}'.format(infected['agent']))
-                    .days_infected
-                    .values[0]
-                )
-                infected_is_contagious = days_infected < pathogen['contagious_period']
-
-                if infected_is_contagious:
-
-                    # Determine if susecptible is immune
-                    immunity = (
-                        self.population
-                        .query('agent == {}'.format(susceptible['agent']))
-                        .immunity
-                        .values[0]
-                    )
-                    susecptible_is_immune = immunity > 0
-
-                    if not susecptible_is_immune:
-                        # Interaction qualified
-                        return True, infected, susceptible
-                    else:
-                        # Interaction disqualified
-                        return False, None, None
-                else:
-                    # Interaction disqualified
-                    return False, None, None
-            else:
-                # Interaction disqualified
-                return False, None, None
-        else:
-            # Interaction disqualified
-            return False, None, None
-
-    def _interact(self, a, b):
-        pathogen = self.pathogen
-
-        # Get agents
-        agent_a, agent_b = self._get_agents(a, b)
-
-        # Qualify interaction
-        interact, infected, susceptible = self._qualify_interaction(agent_a, agent_b)
-
-        # Interact
-        draw = np.random.rand()  # Determines if pathogen spreads
-        infect = draw < pathogen['infection_rate']
-        if interact and infect:
-            self.population.loc[susceptible['agent'], 'state'] = 'inf'
-            self.population.loc[susceptible['agent'], 'infected_by'] = (
-                int(infected['agent'])
-            )
+        return interactions
 
     def interact_all(self):
         verbose = self.verbose
-        logger.debug('- Getting unique interactions.')
-        self._get_unique_connections()
-        connections = self.connections
-        logger.debug('- Running all unique interactions.')
-        for a, b in connections.interaction.values:
-            self._interact(a, b)
+        logger.debug('- Running InteractionEngine with C++ optimizations.')
 
+        interactions = self._init_interactions()
+
+        # Run the Interaction Engine functionality
+        interactions.interact_all()
+        results = interactions.get_population_df()
+        self.population = results
+        if verbose:
+            logger.debug(
+                '- Population states: {}'
+                .format(
+                    self
+                    .population
+                    .state
+                    .value_counts(normalize=True)
+                    .to_dict()
+                )
+            )
+
+
+class InteractionEngineCpp2(InteractionEngineCpp):
+    def _prep_for_cpp(self) -> [dict, dict, dict]:
+        '''Preps data to pass off to cython functions.
+
+        Returns:
+            population: dict
+            connections: dict
+            pathogen: dict
+        '''
+        # TODO: Make sure to change __init__ to encode the keys in
+        #       the pathogen dictionary.
+        population = {
+            'agent': self.population['agent'].tolist(),
+            'state': (
+                self.population['state']
+                .apply(lambda state: state.encode())
+                .tolist(),
+            ),
+            'infected_by': self.population['infected_by'].tolist(),
+            'days_infected': self.population['days_infected'].tolist(),
+            'immunity': self.population['immunity'].tolist(),
+            # Working version
+            # pop_immunity = self.population['immunity'].tolist()
+            # NOTE: I specifically tested, and pop_immunity == popuation('immunity'.get)
+            # This also fails, though if I try to load pop_immunity from the original engine
+        }
+
+        connections = {
+            'agent': self.connections['agent'].tolist(),
+            'connections': self.connections['connections'].tolist(),
+            'num_connections': self.connections['num_connections'].tolist(),
+            'max_connections': self.connections['max_connections'].tolist(),
+        }
+
+        pathogen = {k.encode(): v for k, v in self.pathogen.items()}
+
+        return population, connections, pathogen
+
+    @staticmethod
+    def _set_interactions(population, connections, pathogen):
+        '''Returns the PyInteractions instance with population and
+        connections already set.'''
+        interactions = PyInteractions()
+        # TODO: Update method to accept dictionary
+        interactions.set_population(
+            population.get('agent'),
+            population.get('state'),
+            population.get('infected_by'),
+            population.get('days_infected'),
+            population.get('immunity'),
+        )
+        # TODO: Update method to accept dictionary
+        interactions.set_connections(
+            connections['agent'],
+            connections['connections'],
+            connections['num_connections'],
+            connections['max_connections']
+        )
+
+        interactions.set_pathogen_settings(pathogen)
+
+        return interactions
+
+    def interact_all(self):
+        '''Performs interactions between all agents.'''
+        verbose = self.verbose
+        #logger.debug('- Running InteractionEngine with C++ optimizations.')
+        population, connections, pathogen = self._prep_for_cpp()
+
+        interactions = self._set_interactions(
+            population,
+            connections,
+            pathogen
+        )
+
+        # Run the Interaction Engine functionality
+        interactions.interact_all()
+        self.population = interactions.get_population_df()
         logger.debug(
             '- Population states: {}'
             .format(
@@ -214,17 +180,6 @@ class InteractionEngine():
         )
 
 
-pathogen = {
-    'infection_rate': 0.4,
-    'recovery_rate': 0.1,
-    'death_rate': 0.02,
-    'spontaneous_rate': 0.0,
-    'testing_accuracy': None,
-    'immunity_period': 14,
-    'contagious_period': 10,
-    'incubation_period': 1,
-}
-
 if __name__ == '__main__':
-    print("Hi, I'm the Interaction Engine. I'm not meant to be run directly.")
-    print('To use me, please import InteractionEngine in your script.')
+    print("Hi, I'm the InteractionEngine++. I'm not meant to be run directly.")
+    print('To use me, please import InteractionEngineCpp in your script.')
